refactor : RunningRecord Domain refactoring

+++ backend/dduishu/src/main/java/com/b208/dduishu/domain/runningMate/dto/request/RunningMateInfo.java
@@ -13,9 +13,9 @@ public class RunningMateInfo {
     private String id;
     private String imageUrl;
     private LocalDateTime createdAt;
-    private float averageSpeed;
-    private float totalDistance;
-    private float totalTime;
+    private double averageSpeed;
+    private double totalDistance;
+    private double totalTime;
 
     @Builder
     public RunningMateInfo(RunningMate runningMate) {
