+++ backend/dduishu/src/main/java/com/b208/dduishu/domain/runningRecord/dto/request/RunningRecordDetail.java
@@ -21,13 +21,13 @@ public class RunningRecordDetail {
     private RivalRunningRecordInfo rivalRecord;
     private List<RunningRecordOverallInfo> runningRecordInfos;
     private int totalTime;
-    private float totalDistance;
-    private float averageSpeed;
+    private double totalDistance;
+    private double averageSpeed;
     private int averageCalory;
     private LocalDateTime createdAt;
 
     @Builder
-    public RunningRecordDetail(ObjectId id, UserInfo user, CharacterOverview character, RunningType type, RivalRunningRecordInfo rivalRecord, List<RunningRecordOverallInfo> runningRecordInfos, int totalTime, float totalDistance, float averageSpeed, int averageCalory, LocalDateTime createdAt) {
+    public RunningRecordDetail(ObjectId id, UserInfo user, CharacterOverview character, RunningType type, RivalRunningRecordInfo rivalRecord, List<RunningRecordOverallInfo> runningRecordInfos, int totalTime, double totalDistance, double averageSpeed, int averageCalory, LocalDateTime createdAt) {
         this.id = id.toString();
         this.user = user;
         this.character = character;
