+++ backend/dduishu/src/main/java/com/b208/dduishu/domain/runningRecord/dto/request/RunningRecordInfo.java
@@ -20,11 +20,11 @@ public class RunningRecordInfo {
     private RunningType type;
     private ObjectId rivalRecordId;
     private int totalTime;
-    private float totalDistance;
-    private float averageSpeed;
-    private float averageHeartRate;
+    private double totalDistance;
+    private double averageSpeed;
+    private double averageHeartRate;
     private String averagePace;
-    private float stepCount;
+    private double stepCount;
     //러닝 데이터 받아오기
     // 운동일시, 총 시간, 평균 속력, 이동 거리, 평균 심박수
     // + 위치정보
